Implemented functionality for drawing grids, within the class Board. The class Board draws the grid, and it also draw the symbol needed to differentiate the players.
Implemented pygame GUI functionality in the class Game(). THis is mostly
taken from Py_Chess.
Reworked the check_winner and check_draw by having it change the
self.winner and self.draw to a boolean.
Reworked add_input to attempt to add the given input in, and return a
bool if it succeeds or failed